remove superfluous cast

+++ source/BufferGeometry.hpp
@@ -111,7 +111,7 @@ private:
         int numPointersInThisDimension = StdArrayOperations::productCapped(dimIndex+1, m_dimensionExtents);
 
         for (int index = 0; index < numPointersInThisDimension; ++index) {
-            int nextDimExtent = m_dimensionExtents[static_cast<unsigned>(dimIndex) + 1];
+            int nextDimExtent = m_dimensionExtents[dimIndex + 1];
             int offset = startOfNextDimension + nextDimExtent * index;
             // hook up pointer to element of next dimension
             pointerArray[arrayIndex + index] = reinterpret_cast<T*>(&(pointerArray)[offset]);
